Refactoring the grounder.
- Created NoGoodGenerator and moved functionality there from NaiveGrounder.
- Added necessary interfacing between new NoGoodGenerator and NaiveGrounder.
